datatable implemented on admin side

+++ wbee/users/teacher/texam.php
@@ -41,6 +41,13 @@
 	}
 ?>
 
+<?php 
+$count = 0;
+$sql = "SELECT * FROM exam_category";
+$data = $connect->query($sql);
+$count = $data->rowCount();
+?>
+
 <html>
 <head>
 <title>Organization Dashboard</title>
@@ -48,6 +55,11 @@
     <link rel="stylesheet" href="../../css/style.css">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
 	<link rel="stylesheet" href="tstyle.css">
+
+	<!-- sources for datatable -->
+	<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+	<link rel="stylesheet" type="text/css" href="https://cdn.datatables.net/1.10.21/css/jquery.dataTables.css">	
+	<script type="text/javascript" charset="utf8" src="https://cdn.datatables.net/1.10.21/js/jquery.dataTables.js"></script>
 </head>
 	
 <body style="background-color: orange;">
@@ -64,7 +76,7 @@
 				</div>
 			</div>
 			<div class="col-9">
-				<h2>Examination section</h2>
+				<h5>Examination section</h5>
 				<!-- Form to add exam -->
 				<div class="row">
 						<div class="card" style = "background-color: #ffbf00;">
@@ -108,41 +120,58 @@
   					<h6 style="color:red;">Disabling exam is possible with update option in table!</h6>
 				</div>
 				<div class="row">
-						<?php
-						$count = 0;
-						$sql = "SELECT * FROM exam_category";
-						$data = $connect->query($sql);
-						echo '<table class="table table-bordered">
-						<thead>
-							<tr>
-							<th scope="col">#</th>
-							<th scope="col">Stream</th>
-							<th scope="col">Exam Name</th>
-							<th scope="col">Exam Time</th>
-							<th scope="col">Update</th>
-							<th scope="col">Delete</th>
-							<th scope="col">Status</th>
-							</tr>
-						</thead>';
-						
-						foreach($data as $row)
-								{
-									$count=$count+1;
-									echo ' 
-									<tr>
-										<th scope="row">'.$count.'</th>
+
+					<div class="card" style = "background-color: #ffbf00; width: 100%">
+						<div class="card-header">
+							Students performance
+						</div>
+						<div class="card-body">
+									
+							<div class="table-responsive">  
+								<table id="myTable" class="table table-striped table-bordered">  
+									<thead style="background-color: #292a3e; color: white">  
+										<tr>  
+											<th>#</th>  
+											<th>Stream</th> 
+											<th>Subject</th> 
+											<th>Time</th> 
+											<th>Update</th> 
+											<th>Delete</th>
+											<th>Status</th>   
+										</tr>  
+									</thead>  
+									<?php 
+									$serialNum = 0;
+									if($count == 0){
+										echo 'No list yet';
+									} else 
+									{
+										foreach($data as $row)  
+									{  
+										$serialNum = $serialNum+1;
+										echo '  
+										<tr>  
+										<th scope="row">'.$serialNum.'</th>
 										<td>'.$row["sCategory"].'</td>
 										<td>'.$row["category"].'</td>
 										<td>'.$row["exam_time_in_minutes"].'</td>
 										<td><a href="edit_exam.php?id='.$row["id"].'">Update</a></td>
 										<td><a href="delete.php?id='.$row["id"].'">Delete</a></td>
-										<td id="actionStatus">'.$row["status"].'</td>
-									</tr>';
-								}
-								echo '</table>';
-						?>
+										<td id="actionStatus">'.$row["status"].'</td> 
+										</tr>  
+										';  
+									}
+									}
+										
+									?>  
+								</table>  
+							</div>
+						</div>
+					</div> 
+					
 				</div>
-				</div>	
+				<br><br><br><br>
+			</div>	
 		</div>
 	</div>
 	
@@ -151,8 +180,15 @@
 	</div>
 
 	<!-- JS script for bootstrap -->
-<script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
+
 </body>
 </html>
+<script>  
+ $(document).ready( function () {
+    $('#myTable').DataTable(
+		{
+			
+		}
+	);
+} );
+ </script>
